@@ -1,12 +1,17 @@
 plugins {
     id 'java'
-    id "com.github.johnrengelman.shadow" version "4.0.4"
+    id "com.github.johnrengelman.shadow" version "7.0.0"
+    id("io.papermc.paperweight.userdev")
 }
 
 group 'dev.imabad.mceventsuite'
 version '1.0'
 
-sourceCompatibility = 1.9
+java {
+    toolchain {
+        languageVersion.set(JavaLanguageVersion.of(17))
+    }
+}
 
 repositories {
     mavenLocal()
@@ -19,13 +24,9 @@ repositories {
         name "Bungeecord"
         url "https://oss.sonatype.org/content/repositories/snapshots"
     }
-    maven {
-        name "Paper"
-        url 'https://papermc.io/repo/repository/maven-public/'
-    }
+    maven { url = uri("https://repo.papermc.io/repository/maven-public/") }
     maven { url 'https://jitpack.io' }
     maven { url = "https://maven.enginehub.org/repo/" } // WorldEdit
-    maven { url = "https://mvn.intellectualsites.com/content/groups/public" } // PlotSquared
     maven { url = "https://libraries.minecraft.net"} // Authlib / Minecraft
     maven { url = "https://repo.citizensnpcs.co/"} //Citizens
     maven { url = "https://repo.codemc.io/repository/maven-public/"} //CodeMC
@@ -34,20 +35,23 @@ repositories {
 }
 
 dependencies {
-    testCompile group: 'junit', name: 'junit', version: '4.12'
-    compile 'dev.imabad.mceventsuite:eventcore'
+    paperweightDevelopmentBundle("io.papermc.paper:dev-bundle:1.18.2-R0.1-SNAPSHOT")
+    implementation(platform("com.intellectualsites.bom:bom-1.18.x:1.13"))
+//    testCompile group: 'junit', name: 'junit', version: '4.12'
+    implementation("org.hibernate.javax.persistence:hibernate-jpa-2.1-api:1.0.0.Final")
+    implementation 'dev.imabad.mceventsuite:eventcore'
     compileOnly "com.github.MilkBowl:VaultAPI:1.7"
     compileOnly("com.sk89q.worldedit:worldedit-bukkit:7.2.0-SNAPSHOT")
     compileOnly("com.sk89q.worldguard:worldguard-bukkit:7.0.4-SNAPSHOT")
-    compileOnly("com.plotsquared:PlotSquared-Core:6.0.8-SNAPSHOT")
-    compileOnly("com.plotsquared:PlotSquared-Bukkit:6.0.8-SNAPSHOT") { transitive = false }
+    compileOnly("com.plotsquared:PlotSquared-Core:6.9.4")
+    compileOnly("com.plotsquared:PlotSquared-Bukkit:6.9.4")
     compileOnly("com.mojang:authlib:1.5.21")
-    implementation 'net.kyori:adventure-platform-bukkit:4.0.0-SNAPSHOT'
-    compile ("xyz.xenondevs:particle:1.5.1")
+    implementation 'net.kyori:adventure-platform-bukkit:4.1.1'
+    compileOnly ("xyz.xenondevs:particle:1.5.1")
     compileOnly("net.citizensnpcs:citizens:2.0.27-SNAPSHOT")
     compileOnly 'com.gmail.filoghost.holographicdisplays:holographicdisplays-api:2.4.0'
-    compileOnly 'dev.imabad.mceventsuite:cubedcosmetics:1.0'
+//    compileOnly 'dev.imabad.mceventsuite:cubedcosmetics:1.0'
     compileOnly 'org.geysermc:floodgate-bukkit:1.0-SNAPSHOT'
-    compileOnly 'com.destroystokyo.paper:paper-api:1.16.5-R0.1-SNAPSHOT'
+//    compileOnly 'com.destroystokyo.paper:paper-api:1.18.2-R0.1-SNAPSHOT'
     compileOnly files('jars/paper.jar')
 }
