@@ -10,10 +10,10 @@ import dev.imabad.mceventsuite.spigot.modules.shops.api.ISkullProduct;
 import dev.imabad.mceventsuite.spigot.utils.ItemUtils;
 import dev.imabad.mceventsuite.spigot.utils.SoundHelper;
 import dev.imabad.mceventsuite.spigot.utils.StringUtils;
-import net.minecraft.server.v1_16_R3.PacketPlayOutEntityStatus;
+//import net.minecraft.server.v1_16_R3.PacketPlayOutEntityStatus;
 import org.bukkit.Material;
 import org.bukkit.Sound;
-import org.bukkit.craftbukkit.v1_16_R3.entity.CraftPlayer;
+//import org.bukkit.craftbukkit.v1_16_R3.entity.CraftPlayer;
 import org.bukkit.entity.Player;
 import org.bukkit.event.player.PlayerInteractEvent;
 import org.bukkit.inventory.ItemStack;
@@ -67,9 +67,9 @@ public class CakeProduct implements ISkullProduct {
     SoundHelper.playSoundAtPlayer(player, Sound.ENTITY_GENERIC_EAT);
     EventSpigot.getInstance().getServer().getScheduler().runTaskLaterAsynchronously(EventSpigot.getInstance(), () -> {
       EventSpigot.getInstance().getServer().getScheduler().runTaskLaterAsynchronously(EventSpigot.getInstance(), () -> { SoundHelper.playSoundAtPlayer(player, Sound.ENTITY_PLAYER_BURP); }, 20 *2);
-      CraftPlayer cPlayer = ((CraftPlayer)player);
-      PacketPlayOutEntityStatus eatPacket = new PacketPlayOutEntityStatus(cPlayer.getHandle(), (byte)9);
-      cPlayer.getHandle().playerConnection.sendPacket(eatPacket);
+//      CraftPlayer cPlayer = ((CraftPlayer)player);
+//      PacketPlayOutEntityStatus eatPacket = new PacketPlayOutEntityStatus(cPlayer.getHandle(), (byte)9);
+//      cPlayer.getHandle().playerConnection.sendPacket(eatPacket);
       player.getInventory().remove(playerInteractEvent.getItem());
       EventCore.getInstance().getEventPlayerManager().getPlayer(player.getUniqueId()).ifPresent(eventPlayer -> {
         long lastStarblocks = eventPlayer.getLongProperty("lastStarblocks");
